Allow for child DataViews within DataViews when (de)serialising HOCON.
This allows the ConfigurateTranslator to determine when a DataView should be constructed instead of a map. During serialisation, any child DataViews that are detected will have an extra configuration key/value pair applied. Upon deserialisation, this will be detected and reversed, so a DataView, rather than a Map will be restored. This translator will walk down both maps and lists to ensure that DataViews are correctly constructed. Tests are included to attempt to ensure correctness.

This also allows non-maps to be deserialized, the item in question can be retrieved with the
data query "value" - as there must be a DataView at the empty query point.

* Cleaned up some duplicate methods in ConfigurateTranslator

Fixes SpongePowered/SpongeAPI#1543, fixes #1142